% Generated by roxygen2 (4.0.1): do not edit by hand
\name{add.dates}
\alias{add.dates}
\title{Return CDS dates}
\usage{
add.dates(x, date.var = "date", maturity.var = "maturity",
  tenor.var = "tenor", currency.var = "currency")
}
\arguments{
\item{x}{a dataframe containing all necessary information}

\item{date.var}{character column name of date variable}

\item{maturity.var}{character column name of maturity variable}

\item{tenor.var}{character column name of tenor variable}

\item{currency.var}{character column name of currency variable}
}
\value{
a date frame with 8 variables: step-in date (T+1), value date (T+3 business days),
  start date (accrual begin date), end date (maturity), backstop date (T-60
  day look back from which 'protection' is effective), pen coupon date
  (second to last coupon date)
}
\description{
\code{add.dates} Given a dataframe which contains dates, tenor (or maturity) and/or currency
the function returns appropriate dates used in pricing a CDS contract.
}
\examples{
x1 <- data.frame(date = as.Date("2014-05-07"), tenor = 5)
add.dates(x1)

x2 <- data.frame(date = c(as.Date("2014-05-06"), as.Date("2014-05-07")), tenor = rep(5, 2))
add.dates(x2)

x3 <- data.frame(date = as.Date("2014-05-07"), maturity = as.Date("2019-06-20"))
add.dates(x3)
}
\references{
http://www.cdsmodel.com/cdsmodel/assets/cds-model/docs/c-code%20Key%20Functions-v1.pdf
}

