% Generated by roxygen2 (4.0.1): do not edit by hand
\name{adj.next.bus.day}
\alias{adj.next.bus.day}
\title{Adjust to Next Business Day}
\usage{
adj.next.bus.day(date)
}
\arguments{
\item{date}{of \code{Date} class.}
}
\value{
Date adjusted to the following business day
}
\description{
\code{adj.next.bus.day} get the next business day
following 5D bus day convention.
}

