% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adjusted.Binomial.test.R
\name{Adjusted.Binomial.test}
\alias{Adjusted.Binomial.test}
\title{Adjusted Binomial Test}
\usage{
Adjusted.Binomial.test(rating_scale_data, total_observations, PD, DR,
  confidence_level = 0.9, tail = "one", r = 0.4)
}
\arguments{
\item{rating_scale_data}{A master scale data set needs to be specified.}

\item{total_observations}{Total observations variable needs to be specified}

\item{PD}{PD variable needs to be specified.}

\item{DR}{Default rate variable needs to be specified.}

\item{confidence_level}{A confidence level needs to be set. Default value is 0.90.}

\item{tail}{Tail preference of the test needs to be specified.The default value is "one". The other value could be defined as "two".}

\item{r}{The default correlation between rating grades needs to be specified. The default value is 0.40.}
}
\description{
This function allows to perform adjusted binomial test on master scale data.
}
\examples{
Adjusted.Binomial.test(master_scale_data,"Tot_obs", "PD", "Default_rate", 0.95, "two", 0.30)
}
\keyword{creditR}
