% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adjusted.Herfindahl.Hirschman.Index.R
\name{Adjusted.Herfindahl.Hirschman.Index}
\alias{Adjusted.Herfindahl.Hirschman.Index}
\title{Adjusted Herfindahl-Hirschman Index}
\usage{
Adjusted.Herfindahl.Hirschman.Index(data, total_observations)
}
\arguments{
\item{data}{A data set needs to be specified.}

\item{total_observations}{The total observations variable needs to specified.}
}
\description{
This function allows the calculation of Adjusted-HHI for master scales.
}
\examples{
Adjusted.Herfindahl.Hirschman.Index(master_scale_data, "Tot_obs")
}
\keyword{creditR}
