% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anchor.point.R
\name{Anchor.point}
\alias{Anchor.point}
\title{Anchor Point}
\usage{
Anchor.point(master_scale_data, PD, total_observations, central_tendency,
  upper_green = 1.2, upper_red = 1.3, lower_green = 0.8,
  lower_red = 0.7)
}
\arguments{
\item{master_scale_data}{A master scale data set needs to be specified.}

\item{PD}{A PD variable needs to be specified as string.}

\item{total_observations}{A total observations variable needs to be specified as string.}

\item{upper_green}{An upper green threshold needs to be defined. The default value is 1.2.}

\item{upper_red}{An upper red threshold needs to be defined. The default value is 1.3.}

\item{lower_green}{A lower green threshold needs to be defined. The default value is 0.8.}

\item{lower_red}{A lower red threshold needs to be defined. The default value is 0.7.}

\item{central_tendecy}{A  central tendency needs to be specified for testing.}
}
\description{
This function allows the make Anchor point test for a given master scale data.
}
\examples{
Anchor.point(final_scale, "PD","Tot_obs", 0.6)
}
\keyword{creditR}
