% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_uncertainty.R
\name{apply_uncertainty_deduction}
\alias{apply_uncertainty_deduction}
\title{Use uncertainty multiplier to deduct}
\usage{
apply_uncertainty_deduction(posterior_draws)
}
\arguments{
\item{posterior_draws}{posterior draws for some unit}
}
\value{
the number of credits to mint
}
\description{
Use the posterior distribution of ton year draws from a property or project
and calculate the number of credits to mint based on the uncertainty
multiplier
}
