% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian.calibration.R
\name{bayesian.calibration}
\alias{bayesian.calibration}
\title{Bayesian Calibration}
\usage{
bayesian.calibration(data, average_score, total_observations, PD,
  central_tendency, calibration_data, calibration_data_score)
}
\arguments{
\item{data}{The master scale data needs to be specified.}

\item{average_score}{Average score variable in master scale data needs to be specified.}

\item{total_observations}{Total observations variable in master scale data needs to be specified.}

\item{PD}{PD variable in master scale data needs to be specified.}

\item{central_tendency}{The central tendency which is the calibration target needs to be specified.}

\item{calibration_data}{The scoring model data needs to be specified.}

\item{calibration_data_score}{The score variable in calibration data needs to be specified to apply calibration.}
}
\description{
This function allows to make model calibration for a model with bayesian method.
}
\examples{
bayesian.calibration(master_scale_data,"Score", "Total_obs", "PD", 0.06, model_data, "score")
}
\keyword{creditR}
