% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binomial.test.R
\name{Binomial.test}
\alias{Binomial.test}
\title{Binomial Test}
\usage{
Binomial.test(rating_scale_data, total_observations, PD, DR,
  confidence_level = 0.9, tail = "one")
}
\arguments{
\item{rating_scale_data}{A master scale data set needs to be specified.}

\item{total_observations}{Total observations variable needs to be specified}

\item{PD}{PD variable needs to be specified.}

\item{DR}{Default rate variable needs to be specified.}

\item{confidence_level}{A confidence level needs to be set. Default value is 0.90.}

\item{tail}{Tail preference of the test needs to be specified.The default value is "one". The other value could be defined as "two".}
}
\description{
This function allows to perform binomial test on master scale data.
}
\examples{
Binomial.test(master_scale_data,"Tot_obs", "PD", "Default_rate", 0.95, "two")
}
\keyword{creditR}
