% Generated by roxygen2 (4.0.1): do not edit by hand
\name{build.rates}
\alias{build.rates}
\title{Build rates.RData}
\usage{
build.rates(start = as.Date("2004-01-01"), end = as.Date("2005-01-04"))
}
\arguments{
\item{start}{is the start date of the data frame; it is the earliest day
of interest rate that we want in the data frame}

\item{end}{is the end date of the data frame; it is the lastest day
       of interest rate that we want in the data frame}
}
\value{
a data frame that contains the rates based on the ISDA
        pecifications
}
\description{
\code{build.rates} builds the rates.RData stored in the package. The
usage of rates.RData is explained in the rates.Rd help file. Please
type ?rates to see it.
}
\examples{
\dontrun{
build.rates(start = as.Date("2004-01-01"), end = as.Date("2014-08-23"))
}
}
\seealso{
\link{download.markit} \link{download.FRED} \link{rates}
}

