% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_rates.R
\name{build_rates}
\alias{build_rates}
\title{Build a data frame containing interest rates for CDS pricing}
\usage{
build_rates(start.date, end.date)
}
\arguments{
\item{start.date}{is the start date of the data frame; it is the earliest CDS
pricing date that the user concerns.}

\item{end.date}{is the end date of the data frame; it is the latest CDS 
pricing date that the user concerns.}
}
\value{
a data frame that contains the CDS pricing date, the currency, the 
  interest rate expiry and the interest rate.
}
\description{
\code{build_rates} can create a data frame of interest rates from a start 
date to an end date, which are specified by the user. The interest rates will
be later used in CDS pricing. \code{build_rates} also builds the rates.RData 
in the package.
}
\details{
Since the Markit website has the interest rates back to the 1990s, download_markit 
is responsible for building up all the USD interest rate data frame; for EUR 
and JPY, markit can only get from 2005-01-05 to now. But the biggest 
advantage of using download_markit is that, since the Markit website only lists 
the rate expiries that are actually used for CDS pricing, download_markit can
get the exact type of expiries of rates needed to price CDS. Also, it has 
expiry over 1Y to 30Y. Then we want to get data for EUR and JPY before 2005-01-04 (the limit of
markit). 

Another note is that, the rates on Markit have 
not been adjusted to the previous business day. In other words, the rates 
from both website is the exact rate on that day, rather than on the previous 
business day. But download_markit has adjusted the days 
already for the convenience of CDS pricing, so we don't have to worry here.
}
\examples{
\dontrun{
## Running this example will take more than two hours.
build_rates(start.date = Sys.Date() - 10, end.date = Sys.Date())
}
}
\references{
\url{https://www.spglobal.com/en}
}
\seealso{
\link{download_markit} \link{rates}
}
