% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bus.days.offset}
\alias{bus.days.offset}
\title{\code{bus.days.offset} passes R objects to compiled C code loaded into R to adjust for business days.}
\usage{
bus.days.offset(fromDate, offset, holidays = NULL)
}
\arguments{
\item{fromDate}{starting date}

\item{offset}{function.}

\item{holidays}{holidays is an input for holiday files to adjust to
business day counting. Default is \code{None}.}
}
\description{
\code{bus.days.offset} passes R objects to compiled C code loaded into R to adjust for business days.
}

