% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_uncertainty.R
\name{calc_uncertainty_multiplier}
\alias{calc_uncertainty_multiplier}
\title{Calculate uncertainty mutliplier}
\usage{
calc_uncertainty_multiplier(halfwidth_interval_prop)
}
\arguments{
\item{halfwidth_interval_prop}{the proportional halfwidth interval of the
posterior distribution of tonne years}
}
\value{
the uncertainty multiplier for this distribution
}
\description{
Calculate the uncertainity multiplier based on the logistic uncertainty loss
function that was fit to IPCC data. The logistic uncertainty is developed
and explained in \code{cyclops/vignettes/uncertainty_curve.Rmd}, in that file, we
determine the logistic parameters (\code{b0} and \code{b1}).
}
