% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_uncertainty.R
\name{calculate_uncertainty_deduction}
\alias{calculate_uncertainty_deduction}
\title{Calculate Uncertainty Deduction}
\usage{
calculate_uncertainty_deduction(t0_half_width_ratio, t1_half_width_ratio)
}
\arguments{
\item{t0_half_width_ratio}{range(90\% CI) / 2 / C_t0}

\item{t1_half_width_ratio}{range(90\% CI) / 2 / C_t1}
}
\value{
uncertainty deduction
}
\description{
Calculate the project uncertainty deduction fom the T0 and T1 uncertainty estimates
}
