% Generated by roxygen2 (4.0.1): do not edit by hand
\name{call.ISDA}
\alias{call.ISDA}
\title{call ISDA c function}
\usage{
call.ISDA(name, x = x, ratesInfo = ratesInfo, i = i)
}
\arguments{
\item{name}{character function name within which call.ISDA is called}

\item{x}{dataframe which contains relevant dates and convention info}

\item{ratesInfo}{dataframe which contains relevant rates data}

\item{i}{integer which indicates which iteration of the loop we are currently at}
}
\value{
a numeric value which is the difference between the new upfront and the old one
}
\description{
\code{call.ISDA} call ISDA function
}

