% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_ISDA.R
\name{call_ISDA}
\alias{call_ISDA}
\title{call ISDA c function}
\usage{
call_ISDA(x, name, rates.info)
}
\arguments{
\item{x}{dataframe which contains relevant dates and convention info}

\item{name}{character function name within which call_ISDA is called}

\item{rates.info}{dataframe which contains relevant rates data}
}
\value{
a numeric value which is the difference between the new upfront and 
  the old one
}
\description{
\code{call_ISDA} call ISDA function
}

