% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CDS-package.R, R/CDS.R
\docType{package}
\name{CDS}
\alias{CDS}
\alias{CDS-package}
\title{The CDS package.}
\usage{
CDS(name = NULL, contract = "SNAC", RED = NULL, date = Sys.Date(),
  spread = NULL, maturity = NULL, tenor = NULL, coupon = 100,
  recovery = 0.4, currency = "USD", notional = 1e+07)
}
\arguments{
\item{name}{is the name of the reference entity. Optional.}

\item{contract}{is the contract type, default SNAC}

\item{RED}{alphanumeric code assigned to the reference entity. Optional.}

\item{date}{is when the trade is executed, denoted as T. Default is
\code{Sys.Date}. The date format should be in "YYYY-MM-DD".}

\item{spread}{CDS par spread in bps.}

\item{maturity}{date of the CDS contract.}

\item{tenor}{of contract. By default is set as 5}

\item{coupon}{quoted in bps. It specifies the payment amount from the
protection buyer to the seller on a regular basis. The default is 100 bps.}

\item{recovery}{in decimal. Default is 0.4.}

\item{currency}{in which CDS is denominated.}

\item{notional}{is the amount of the underlying asset on which the payments
  are based. Default is 1e7, i.e. 10MM.}
}
\value{
a \code{CDS} class object including the input informtion on the
  contract as well as the valuation results of the contract.
}
\description{
\code{creditr} package provides useful tools for pricing credit
default swaps (CDS). It enables CDS class object which has slots as
name, contract, RED, date, spread, maturity, teno, coupon, recovery,
currency, notional, principal, accrual, pd, price, upfront,
spread.DV01, IR.DV01 and rec.risk.01, with S4 methods like update,
show and summary. It also supports data frame input and is able to
provide convenient calculation of key CDS statistics through functions
like CS10, IR.DV01, rec.risk.01 and spread.DV01. Of other major
functions, spread.to.upfront and upfront.to.spread are designed to
compute one of spread and upfront given the other; spread.to.pd and
pd.to.spread, similarly, can calculate one of spread and probability
of default given the other; add.dates and add.conventions compute a
series of dates information and accounting conventions related to CDS
pricing. Finally, get.rates and build.rates facilitates direct
fetching of relevant interest rates from online sources. Thanks to
ISDA Standard Model's Open Source license, we are able to create this
package for R users. You can find the Open Source licence of ISDA
Standard Model at "http://www.cdsmodel.com/cdsmodel/cds-disclaimer.html?"

Build a \code{CDS} class object given the input about a CDS contract.
}
\examples{
x <- CDS(date = as.Date("2014-05-07"), tenor = 5, spread = 50, coupon = 100)
}

