% Generated by roxygen2 (4.0.1): do not edit by hand
\name{check.date}
\alias{check.date}
\title{\code{check.date} for testing if the date entered is valid. This
function is used in CDS.R, CS10, get.rates.DF, IR.DV01, rec.risk
spread.DV.01, and upfront.}
\usage{
check.date(date)
}
\arguments{
\item{date}{entered}
}
\value{
TRUE if date is valid, FALSE if date is in future, and
"Invalid date format. Must be YYYY-MM-DD" if format is wrong
}
\description{
\code{check.date} for testing if the date entered is valid. This
function is used in CDS.R, CS10, get.rates.DF, IR.DV01, rec.risk
spread.DV.01, and upfront.
}

