% Generated by roxygen2 (4.0.1): do not edit by hand
\name{check.rates.dates}
\alias{check.rates.dates}
\title{Function that makes sure the dates in data frame x
being used by the upfrontdf function are also present in the
rates data frame being entered into the upfrontdf function (if
it is not contained in the rates data frame, the upfrontdf
function will be unable to return the upfront value)
match with the rates on markit.com for that specific date.}
\usage{
check.rates.dates(x, rates)
}
\arguments{
\item{x}{data frame which contains rates for a specific date}

\item{rates}{data frame which contains rates and dates}
}
\value{
vector containing true or false to indicate whether
that specific trade date in X is contained in rates
}
\description{
Function that makes sure the dates in data frame x
being used by the upfrontdf function are also present in the
rates data frame being entered into the upfrontdf function (if
it is not contained in the rates data frame, the upfrontdf
function will be unable to return the upfront value)
match with the rates on markit.com for that specific date.
}

