% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisquare.test.R
\name{chisquare.test}
\alias{chisquare.test}
\title{Chi Square Test}
\usage{
chisquare.test(data, PD, observed_bad, total_observations,
  confidence_level = 0.95)
}
\arguments{
\item{data}{A master scale data set needs to be specified.}

\item{PD}{PD variable needs to be specified.}

\item{observed_bad}{Observed defaults needs to be specified.}

\item{total_observations}{Total observations variable needs to be specified}

\item{confidence_level}{A confidence level needs to be set. Default value is 0.95.}
}
\description{
This function allows to perform chi square test on master scale data.
}
\examples{
chisquare.test(master_scale_data, "PD","Bad_obs","Tot_obs",0.90)
}
\keyword{creditR}
