% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creditr.R
\name{compute_raw_credits}
\alias{compute_raw_credits}
\title{Composite function to compute and verify boundary of credits}
\usage{
compute_raw_credits(
  t0_tCO2,
  annual_growth_rate,
  baseline_intensity,
  loss_intensity,
  supersection,
  annual_discount_rate = 0.03,
  leakage = 0.2
)
}
\arguments{
\item{t0_tCO2}{[0, inf) metric tonnes of CO2 at day 0}

\item{annual_growth_rate}{[0, inf) average annual growth of carbon}

\item{baseline_intensity}{[0-1] proportion of carbon harvested under BAU}

\item{loss_intensity}{[0-1] estimated carbon harvested at end of performance period}

\item{supersection}{string; USFS supersection}

\item{annual_discount_rate}{numeric [0, inf); annualized discount rate for carbon deferred this year}

\item{leakage}{numeric [0-1]; deduction to be applied for leakage}
}
\description{
Composite function to compute and verify boundary of credits
}
