% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.cluster.R
\name{correlation.cluster}
\alias{correlation.cluster}
\title{Average Correlations of Clusters}
\usage{
correlation.cluster(data, clustering_data, variables, clusters)
}
\arguments{
\item{data}{A raw data set needs to be specified.}

\item{clustering_data}{Output of the clustering function needs to be specified.}

\item{variables}{The column name of "variables" of clustering function output needs to be specified.}

\item{clusters}{The column name of "clusters" of clustering function output needs to be specified.}
}
\description{
This function allows the calculation of correlations for clusters. This function should be used with clustering function.
}
\examples{
correlation.cluster(my_data,clustering_output, "variable_names","Groups")
}
\keyword{creditR}
