% Generated by roxygen2 (4.0.1): do not edit by hand
\name{CS10.df}
\alias{CS10.df}
\title{function that takes a data frame containing the trade date, spread, coupon and tenor to return
the corresponding CS 10 column with the rest of the data frame}
\usage{
CS10.df(x)
}
\arguments{
\item{x}{data frame containing the trade date, spread, coupon and tenor}
}
\value{
original data frame with a column containing the corresponding CS10 values
}
\description{
function that takes a data frame containing the trade date, spread, coupon and tenor to return
the corresponding CS 10 column with the rest of the data frame
}

