% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creditr.R
\name{D}
\alias{D}
\title{Decay / discount function}
\usage{
D(lambda, t)
}
\arguments{
\item{lambda}{numeric; decay / discount coefficient. For annual X\% decay /
discounting \code{lambda = -log(1 - X/100)}}

\item{t}{numeric; time in years}
}
\description{
Decay / discount function
}
