% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{decay_coefs}
\alias{decay_coefs}
\title{decay_coefs}
\format{
A tibble with 9 rows and 2 variables:
\describe{
\item{product}{harvest product}
\item{lambda}{decay parameter}
}
}
\source{
data-raw/hwp.R
}
\usage{
decay_coefs
}
\description{
Decay rate parameters by harvest product
}
\keyword{datasets}
