% Generated by roxygen2 (4.0.1): do not edit by hand
\name{default.expo}
\alias{default.expo}
\title{\code{default.expo} to calculate the default exposure of a CDS contract based on the
formula: Default Exposure: (1-Recovery Rate)*Notional - Principal}
\usage{
<<<<<<< HEAD:pkg/man/defaultExpo.Rd
defaultExpo(recoveryRate = 0.4, notional = 10000000, principal)
=======
default.expo(recoveryRate = 0.4, notional = 1e+07, principal)
>>>>>>> e401d0df7eb96bb0e12a5c838a70b8ac67383b22:pkg/man/default.expo.Rd
}
\arguments{
\item{recoveryRate}{in decimal. Default is 0.4}

\item{notional}{is the notional amount of the CDS contract. Default
is 10MM.}

\item{principal}{is the principal from the CDS contract.}
}
\value{
a numeric indicating the amount in the event of a default
the following day.
}
\description{
\code{default.expo} to calculate the default exposure of a CDS contract based on the
formula: Default Exposure: (1-Recovery Rate)*Notional - Principal
}

