% Generated by roxygen2 (4.0.1): do not edit by hand
\name{default.prob}
\alias{default.prob}
\title{Approximate the default probability at time t given the parSpread}
\usage{
default.prob(parSpread, t, recoveryRate = 0.4)
}
\arguments{
\item{parSpread}{in bps}

\item{t}{in years}

\item{recoveryRate}{in decimal. Default is 0.4.}
}
\description{
Approximate the default probability at time t given the parSpread
}

