% Generated by roxygen2 (4.0.1): do not edit by hand
\name{download.FRED}
\alias{download.FRED}
\title{Get Rates from FRED}
\usage{
download.FRED(start = as.Date("2004-01-01"), end = as.Date("2005-01-04"),
  currency = "JPY")
}
\arguments{
\item{date}{is the Trade date.
The rates for a trade date T are published on T-1 weekday.
This date refers to the day on which we want the CDS to be
priced, not the date for the interest rates as the interest
rates will be used is the day before the trade date. Eg. If we
are trying to find the rates used to price a CDS on 2014-04-22,
it will return the rates of 2014-04-21}

\item{currency}{is the three-letter currency code. As of now, it works
       for USD, EUR, and JPY. The default is JPY.}
}
\value{
a data frame that contains the rates based on the ISDA
        pecifications
}
\description{
\code{download.FRED} returns the deposits and swap rates for the day
input, along with the date conventions for that specific currency.
The source is FRED.
}
\examples{
\dontrun{
download.FRED(start = as.Date("2003-12-31"), end = as.Date("2005-01-04"),
              currency = "JPY")
}
}

