% Generated by roxygen2 (4.0.1): do not edit by hand
\name{download.rates}
\alias{download.rates}
\title{Download Rates from the Internet}
\usage{
download.rates(URL, verbose = FALSE)
}
\arguments{
\item{URL}{is the link containing the rates.}

\item{verbose}{option. Default \code{FALSE}.}
}
\value{
a xml file crawled from the \code{URL}.
}
\description{
\code{download.rates} download the rates zip file from a given URL. Unzip and parse the
XML
}

