% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creditr.R
\name{E}
\alias{E}
\title{Emission function}
\usage{
E(lambda, t)
}
\arguments{
\item{lambda}{numeric; emission coefficient. For annual X\% emission use
\code{lambda = -log(1 - X/100)}}

\item{t}{numeric; time in years}
}
\description{
Emission function
}
