% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creditr.R
\name{E_prime}
\alias{E_prime}
\title{Instantaneous emission function}
\usage{
E_prime(lambda, t)
}
\arguments{
\item{lambda}{numeric; decay coefficient. For annual X\% emission use
\code{lambda = -log(1 - X/100)}}

\item{t}{numeric; time in years}
}
\description{
The derivative of \code{\link{E}}
}
