% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creditr.R
\name{F}
\alias{F}
\title{Discounted instantaneous emission function}
\usage{
F(lambda, rho, t, start = 0)
}
\arguments{
\item{lambda}{numeric; decay coefficient. For annual X\% decay
\code{lambda = -log(1 - X/100)}}

\item{rho}{numeric; discount coefficient. For annual X\% discount
\code{rho = -log(1 - X/100)}}

\item{t}{numeric; time in years}

\item{start}{numeric; time in years at which the emission begins. Acts as
an offset for emissions that don't begin until a future date.}
}
\description{
Discounted instantaneous emission function
}
