% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creditr.R
\name{G_prime}
\alias{G_prime}
\title{Instantaneous growth function}
\usage{
G_prime(gamma, t)
}
\arguments{
\item{gamma}{numeric; growth coefficient. For annual X\% growth, use
\code{gamma = log(1 + X/100)}}

\item{t}{numeric; time in years}
}
\description{
Instantaneous growth function
}
