% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get.date}
\alias{get.date}
\title{Return CDS dates}
\usage{
get.date(date, maturity = NULL, tenor = NULL, currency = "USD")
}
\arguments{
\item{date}{the date for which we are calculating the relevant dates. Must be of class "date".}

\item{maturity}{date of maturity of the CDS contract. Default is NULL. Must be of class "date"}

\item{tenor}{numeric tenor of the CDS contract in format like 5, 3. Default
is NULL. Must enter either maturity or tenor, but not both.}

\item{currency}{of the contract for which we are calculating the relevant dates. The currency affects
the calculation of the baseDate as in the case of JPY denominated contracts, the baseDate cannot
lie on any of the specific Japanese holidays.}
}
\value{
a date frame with 8 variables: step-in date (T+1), value date (T+3 business days),
  start date (accrual begin date), end date (maturity), backstop date (T-60
  day look back from which 'protection' is effective), pen coupon date
  (second to last coupon date)
}
\description{
\code{get.date} returns appropriate dates used in pricing a CDS contract.
}
\examples{
get.date(as.Date("2014-05-07"), tenor = 5, maturity = NULL)
}
\references{
http://www.cdsmodel.com/cdsmodel/assets/cds-model/docs/c-code%20Key%20Functions-v1.pdf
}

