% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getDates}
\alias{getDates}
\title{The function gets relevant coupon dates for a CDS contract.}
\usage{
getDates(TDate, maturity = NULL, tenor = NULL)
}
\arguments{
\item{TDate}{the trade date}

\item{maturity}{date of maturity of the CDS contract. Default is NULL.}

\item{tenor}{character tenor of the CDS contract in format like 5Y. Default
is NULL. Must enter either maturity or tenor, but not both.}
}
\value{
a date frame with step-in date (T+1), value date (T+3 business days),
  start date (accrual begin date), end date (maturity), backstop date (T-60
  day look back from which 'protection' is effective), pen coupon date
  (second to last coupon date)
}
\description{
The function gets relevant coupon dates for a CDS contract.
}
\examples{
getDates(as.Date("2014-05-07"), tenor = "5Y", maturity = NULL)
}

