% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gini.univariate.R
\name{Gini.univariate}
\alias{Gini.univariate}
\title{Univariate Gini}
\usage{
Gini.univariate(data, default_flag, variable)
}
\arguments{
\item{data}{A data set needs to be specified.}

\item{default_flag}{The default flag need to specified as string.}

\item{variable}{The name of the variable need to specified to calculate Gini value.}
}
\description{
This function performs the calculation of the Gini from the estimated values calculated by logistic regression of a variable. It should be noted that the Gini values calculated by logistic regression.
}
\examples{
default_f <- c('1','0','0', '1','1','0','0','1','1')
birth_year <- c(1980, 1985, 1971,1971,1985,1971,1980,1980,1985)
job <- c(1,1,2, 2,2,3,3,2,3)
example_data <- data.frame(default_f,birth_year,job)
Gini.univariate(example_data, "default_f", "birth_year")
}
\keyword{creditR}
