% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{growth_rates}
\alias{growth_rates}
\title{growth_rates}
\format{
A tibble with 33 rows and 3 variables:
\describe{
\item{fldgrpcd}{USFS field group code}
\item{mean_annual_growth_carbon_tons_per_ac_pct}{population estimated growth rates from rFIA}
\item{s_multiplier}{sequestration multiplier}
}
}
\source{
data-raw/growth_rates.R
}
\usage{
growth_rates
}
\description{
USFS field group-associated growth rates/sequestration multipliers
}
\keyword{datasets}
