% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creditr.R
\name{H}
\alias{H}
\title{Discounted instantaneous growth function}
\usage{
H(gamma, rho, t)
}
\arguments{
\item{gamma}{numeric; growth coefficient. For annual X\% growth
\code{gamma = log(1 + X/100)}}

\item{rho}{numeric; discount coefficient. For annual X\% discount
\code{rho = -log(1 - X/100)}}

\item{t}{numeric; time in years}
}
\description{
Discounted instantaneous growth function
}
