% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Herfindahl.Hirschman.Index.R
\name{Herfindahl.Hirschman.Index}
\alias{Herfindahl.Hirschman.Index}
\title{Herfindahl-Hirschman Index}
\usage{
Herfindahl.Hirschman.Index(data, total_observations)
}
\arguments{
\item{data}{A data set needs to be specified.}

\item{total_observations}{The total observations variable needs to specified.}
}
\description{
This function allows the calculation of HHI for master scales.
}
\examples{
Herfindahl.Hirschman.Index(master_scale_data, "Tot_obs")
}
\keyword{creditR}
