% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hwp_decay}
\alias{hwp_decay}
\title{hwp_decay}
\format{
A tibble with 196 rows and 6 variables:
\describe{
\item{year}{decay year}
\item{product}{harvested wood product}
\item{prop_remaining}{
proportion of carbon remaining in \code{product} at \code{year}
}
\item{prop_emitted}{
cumulative proportion of carbon emitted into the atmosphere for \code{product}
in \code{year}; computed as 1 - \code{prop_remaining}
}
}
}
\source{
data-raw/hwp.R
}
\usage{
hwp_decay
}
\description{
Decay rate information for classes of harvested wood products.
}
\keyword{datasets}
