% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hwp_mix}
\alias{hwp_mix}
\title{hwp_mix}
\format{
A tibble with 658 rows and 3 variables:
\describe{
\item{supersection}{USFS supersection}
\item{product}{harvested wood product}
\item{mix_prop}{
proportion of harvest allocated to \code{product} in a given supersection
}
}
}
\source{
data-raw/hwp.R
}
\usage{
hwp_mix
}
\description{
Typical harvested wood product mixtures by supersection.
}
\keyword{datasets}
