% Generated by roxygen2 (4.0.1): do not edit by hand
\name{implied.RR}
\alias{implied.RR}
\title{Calculates Implied Recovery Rate}
\usage{
implied.RR(data, spread.var = "spread", pd.var = "pd",
  end.date.var = "end.date", date.var = "date")
}
\arguments{
\item{data}{dataframe containing the 1. probability of default (pd) over a a
certain tenor period, 2. id and 3. spread}

\item{spread.var}{name of the column containing the spread}

\item{end.date.var}{name of the column containing the end dates or maturity
dates. This is when the contract expires and protection ends. Any default
after this date does not trigger a payment.}

\item{date.var}{name of the column containing the trade dates.}

\item{pd.var}{name of the column containing the probability of default rates.}
}
\value{
implied recovery rate in percentage based on the general
  approximation for a probability of default in the Bloomberg manual. The
  actual calculation uses a complicated bootstrapping process, so the results
  may be marginally different.
}
\description{
\code{implied.RR} that calculates the recovery rate implied by the CDS spread
and probability of default (pd) by using the ISDA model. This takes a data
frame of inputs and returns a vector of the same length.
}

