% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{IR.DV.01-method}
\alias{IR.DV.01,CDS-method}
\alias{IR.DV.01-method}
\title{The IRDV01 method for CDS class. Calculate the IRDV01 from
conventional spread}
\usage{
\S4method{IR.DV.01}{CDS}(object)
}
\arguments{
\item{object}{the input CDS class object}
}
\description{
The IRDV01 method for CDS class. Calculate the IRDV01 from
conventional spread
}

