% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IV.calc.data.R
\name{IV.calc.data}
\alias{IV.calc.data}
\title{Calculate Information Value for a Given Data}
\usage{
IV.calc.data(data, default_flag)
}
\arguments{
\item{data}{A data set needs to be specified.}

\item{default_flag}{The default flag need to specified as string.}
}
\description{
This function allows the calculation of the IV for a given data.
}
\examples{
default_f <- c('1','0','0', '1','1','0','0','1','1')
birth_year <- c(1980, 1985, 1971,1971,1985,1971,1980,1980,1985)
job <- c(1,1,2, 2,2,3,3,2,3)
example_data <- data.frame(default_f,birth_year,job)
IV.calc.data(example_data, "default_f")
}
\keyword{creditR}
