% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IV_elimination.R
\name{IV_elimination}
\alias{IV_elimination}
\title{Information Value Elimination}
\usage{
IV_elimination(data, default_flag, iv_threshold)
}
\arguments{
\item{data}{A data set needs to be defined.}

\item{default_flag}{Default flag must be specified as a string.}

\item{iv_threshold}{The IV threshold must be specified in order to make an elimination.}
}
\description{
This function allows you to eliminate variables which have information value less than a given threshold for a given data set.
}
\examples{
default_f <- c('1','0','0', '1','1','0','0','1','1')
birth_year <- c(1980, 1985, 1971,1971,1985,1971,1980,1980,1985)
job <- c(1,1,2, 2,2,3,3,2,3)
example_data <- data.frame(default_f,birth_year,job)
IV_elimination(example_data,"default_f" ,0.50)
}
\keyword{creditR}
