% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k.fold.cross.validation.glm.R
\name{k.fold.cross.validation.glm}
\alias{k.fold.cross.validation.glm}
\title{K Fold Cross Validation Gini}
\usage{
k.fold.cross.validation.glm(model_data, default_flag, folds, seed_value)
}
\arguments{
\item{model_data}{A data set needs to be specified.}

\item{default_flag}{The default flag needs to specified as string.}

\item{folds}{The number of folds desired needs to be specified.}

\item{seed_value}{A seed value needs to specified for replicability.}
}
\description{
This function allows to create k fold cross validation data sets.
}
\examples{
default_f <- c('1','0','0', '1','1','0','0','1','1')
birth_year <- c(1980, 1985, 1971,1971,1985,1971,1980,1980,1985)
job <- c(1,1,2, 2,2,3,3,2,3)
example_data <- data.frame(default_f,birth_year,job)
k.fold.cross.validation.glm(example_data, "default_f",10, 1)
}
\keyword{creditR}
