% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kolmogorov.Smirnov.R
\name{Kolmogorov.Smirnov}
\alias{Kolmogorov.Smirnov}
\title{Calculate Kolmogorov-Smirnov Statistic for a Given Scoring Data}
\usage{
Kolmogorov.Smirnov(data, default_flag, PD)
}
\arguments{
\item{data}{A data set needs to be specified.}

\item{default_flag}{The default flag needs to specified as string.}

\item{PD}{The PD variable needs to be specified.}
}
\description{
This function allows the calculation of the KS for a given scoring data.
}
\examples{
Kolmogorov.Smirnov(model_data, "Default_f", "PD")
}
\keyword{creditR}
