% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max.gini.model.R
\name{max.gini.model}
\alias{max.gini.model}
\title{Maximum Gini Model}
\usage{
\method{max}{gini.model}(data, default_flag, seed_value = 1)
}
\arguments{
\item{data}{A data set needs to be specified.}

\item{default_flag}{The default flag need to specified as string.}

\item{seed_value}{A seed value need to specified for replicability.}
}
\description{
This function allows the finding the model which gives the maximum gini value. Statistical requirements will not be provided. Can only be used to give an inference.
}
\examples{
default_f <- c('1','0','0', '1','1','0','0','1','1')
birth_year <- c(1980, 1985, 1971,1971,1985,1971,1980,1980,1985)
job <- c(1,1,2, 2,2,3,3,2,3)
example_data <- data.frame(default_f,birth_year,job)
max.gini.model(example_data, "default_f",10)
}
\keyword{creditR}
