% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_elimination.R
\name{missing_elimination}
\alias{missing_elimination}
\title{Missing Elimination}
\usage{
missing_elimination(data, missing_ratio_threshold)
}
\arguments{
\item{data}{A data set needs to be defined.}

\item{missing_ratio_threshold}{The missing ratio threshold must be specified in order to make an elimination.}
}
\description{
This function allows you to eliminate variables which have missing ratios greater than a given threshold for a given data set.
}
\examples{
name <- c('John Doe','Peter Gynn','Jolie Hope')
birth_year <- c(1980, 1985, 1971)
salary <- c(20000,NA,10000)
example_data <- data.frame(name,birth_year,salary)
missing_elimination(example_data, 0.10)
}
\keyword{creditR}
