% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_ratio.R
\name{missing_ratio}
\alias{missing_ratio}
\title{Missing Ratio}
\usage{
missing_ratio(data)
}
\arguments{
\item{data}{A data set needs to be defined.}
}
\description{
This function allows you to calculate the missing ratios of variables for a given data set.
}
\examples{
name <- c('John Doe','Peter Gynn','Jolie Hope')
birth_year <- c(1980, 1985, 1971)
salary <- c(20000,NA,10000)
example_data <- data.frame(name,birth_year,salary)
missing_ratio(example_data)
}
\keyword{creditR}
