% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_checker.R
\name{na_checker}
\alias{na_checker}
\title{NA Checker}
\usage{
na_checker(data)
}
\arguments{
\item{data}{A data set needs to be defined.}
}
\description{
This function controls the presence of NA values in variables for a given data set. For the results returned as TRUE, the corresponding variable has NA observation/observations.
}
\examples{
name <- c('John Doe','Peter Gynn','Jolie Hope')
birth_year <- c(1980, 1985, 1971)
salary <- c(20000,NA,10000)
example_data <- data.frame(name,birth_year,salary)
na_checker(example_data)
}
\keyword{creditR}
