% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pd.spread}
\alias{pd.spread}
\title{\code{pd.spread} to calculate spread using the probability of default, time and recovery rate}
\usage{
pd.spread(PD, tenor, TDate = Sys.Date(), Recovery = 0.4)
}
\arguments{
\item{Recovery}{rate in decimals}

\item{PD}{probability of default in decimals}

\item{tenor}{in years}

\item{TDate}{trade date. By default is the current date}
}
\value{
spread in basis points, calculated by inverting the formula for probability
of default given in the Bloomberg Manual
}
\description{
\code{pd.spread} to calculate spread using the probability of default, time and recovery rate
}

