% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pd.to.spread}
\alias{pd.to.spread}
\title{Calculate spread with Default Probability}
\usage{
pd.to.spread(x, recovery.var = "recovery", tenor.var = "tenor",
  date.var = "date", pd.var = "PD")
}
\arguments{
\item{x}{dataset containing the recovery rate, tenor (in years), probability
of default and trade date.}

\item{recovery.var}{name of the column containing the recovery rate in
decimals.}

\item{pd.var}{name of the column containing the probability of default in
decimals.}

\item{tenor.var}{name of the column containing the tenor in years.}

\item{date.var}{name of the column containing the trade date.}
}
\value{
vector containing the spread values in basis points, calculated by
  inverting the formula for probability of default given in the Bloomberg
  Manual
}
\description{
\code{pd.to.spread} to calculate spread using the probability of default,
tenor and recovery rate.
}

