% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSI.calc.data.R
\name{PSI.calc.data}
\alias{PSI.calc.data}
\title{Calculate Population Stability Index for a Given Data}
\usage{
PSI.calc.data(main_data, second_data, bin = 10)
}
\arguments{
\item{main_data}{The main data set of a measurement, should be a factor or numeric}

\item{bin}{The desired number of bins should be specified. Default value is 10.}

\item{secon_data}{The second data set of a measurement, should be a factor or numeric}
}
\description{
This function allows the calculation of the PSI for a given data.
}
\examples{
data("iris")
train <- sample(nrow(iris), nrow(iris) * .7)
test <- iris[-train]
p <- PSI.calc(train, test)
p

}
\keyword{creditR}
