% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{rec.risk.01-method}
\alias{rec.risk.01,CDS-method}
\alias{rec.risk.01-method}
\title{Calculate the amount of change in upfront when there is a 1\%
increase in recovery rate.}
\usage{
\S4method{rec.risk.01}{CDS}(object)
}
\arguments{
\item{object}{the input \code{CDS} class object}
}
\description{
Calculate the amount of change in upfront when there is a 1\%
increase in recovery rate.
}

