% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.calibration.R
\name{regression.calibration}
\alias{regression.calibration}
\title{Regression Calibration}
\usage{
regression.calibration(model, calibration_data, default_flag)
}
\arguments{
\item{model}{The scoring model needs to be specified.}

\item{calibration_data}{A calibration data needs to be specified.}

\item{default_flag}{The default flag need to specified as string.}
}
\description{
This function allows to make model calibration for a model with logistic regression.
}
\examples{
regression.calibration(scoring_model, calibration_sample, "default_f")
}
\keyword{creditR}
