% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaled.score.R
\name{scaled.score}
\alias{scaled.score}
\title{Scaled Score}
\usage{
scaled.score(data, PD, ceiling_score, increase)
}
\arguments{
\item{data}{A data set needs to be specified including PD variable.}

\item{PD}{PD variable need to be specified.}

\item{ceiling_score, }{A ceiling score needs to be specified for transformation.}

\item{increase, }{An increase level needs to be specified for transformation.}
}
\description{
This function allows the creation of scaled score for a given data.
}
\examples{
default_f <- c('1','0','0', '1','1','0','0','1','1')
birth_year <- c(1980, 1985, 1971,1971,1985,1971,1980,1980,1985)
PD <- c(0.1,0.12, 0.2, 0.23, 0.28,0.33,0.39,0.45, O.54)
example_data <- data.frame(default_f,birth_year,PD)
scaled.score(example_data, "PD", 1000, 15)
}
\keyword{creditR}
