% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/separate.YMD.R
\name{separate.YMD}
\alias{separate.YMD}
\title{Separate Year/Month/Day}
\usage{
separate.YMD(d)
}
\arguments{
\item{d}{is an input date.}
}
\value{
an array contains year, month, date of the input date
\code{d}.
}
\description{
\code{separate.YMD} contains helper functions to separate an input date into
 year, month, and day.
}

