% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_YMD.R
\name{separate_YMD}
\alias{separate_YMD}
\title{Separate Year/Month/Day}
\usage{
separate_YMD(d)
}
\arguments{
\item{d}{is an input date.}
}
\value{
an array contains year, month, date of the input date
\code{d}.
}
\description{
\code{separate_YMD} contains helper functions to separate an input date into
 year, month, and day.
}

