% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{spread.DV.01-method}
\alias{spread.DV.01,CDS-method}
\alias{spread.DV.01-method}
\title{The spread.DV.01 method for CDS class.}
\usage{
\S4method{spread.DV.01}{CDS}(object)
}
\arguments{
\item{object}{the input \code{CDS} class object}
}
\description{
Calculate the \code{spread.DV.01} from conventional spread
}

