% Generated by roxygen2 (4.0.1): do not edit by hand
\name{spread}
\alias{spread}
\title{\code{spread} to calculate conventional spread using the upfront or ptsUpfront values}
\usage{
spread(x, currency.var = "currency", date.var = "date",
  coupon.var = "coupon", tenor.var = "tenor", maturity.var = "maturity",
  RR.var = "recovery.rate", upfront.var = "upfront",
  ptsUpfront.var = "ptsUpfront", isPriceClean = FALSE, notional = 1e+07,
  freqCDS = "Q", stubCDS = "F", payAccruedAtStart = FALSE,
  payAccruedOnDefault = TRUE)
}
\arguments{
\item{x}{data frame which contains information of date, tenor, coupon, upfront, etc.}

\item{currency.var}{}

\item{date.var}{}

\item{coupon.var}{}

\item{tenor.var}{}

\item{maturity.var}{}

\item{RR.var}{}

\item{upfront.var}{}

\item{ptsUpfront.var}{}

\item{isPriceClean}{}

\item{notional}{}

\item{freqCDS}{date interval of the CDS contract.}

\item{stubCDS}{is a character indicating the presence of a stub.}

\item{payAccruedAtStart}{whether pay at start date the accrual amount}

\item{payAccruedOnDefault}{}
}
\value{
a numeric indicating the spread.
}
\description{
\code{spread} to calculate conventional spread using the upfront or ptsUpfront values
}

