% Generated by roxygen2 (4.0.1): do not edit by hand
\name{spread.to.pd}
\alias{spread.to.pd}
\title{Calcualte Default Probability with Spread}
\usage{
spread.to.pd(spread, time, recovery.rate = 0.4)
}
\arguments{
\item{spread}{in bps}

\item{time}{in years}

\item{recovery.rate}{in decimal. Default is 0.4.}
}
\value{
default probability in decimals
}
\description{
\code{spread.to.pd} approximates the default probability at time given the spread
}

