% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSI.calc.data.R
\name{SSI.calc.data}
\alias{SSI.calc.data}
\title{Calcuate the System Stability Index for a Given Data}
\usage{
SSI.calc.data(main_data, second_data, default_flag)
}
\arguments{
\item{main_data}{A main data set needs to be specified.}

\item{second_data}{A second data set needs to be specified.}

\item{default_flag}{The default flag need to specified as string to exlude from the data.}
}
\description{
This function allows the calculation of the SSI for a given data.
}
\examples{
SSI.calc.data(data1, data2, "default_f")
}
\keyword{creditR}
