% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSI.calc.R
\name{SSI.calc}
\alias{SSI.calc}
\title{Calcuate the System Stability Index for a Given Variable}
\usage{
SSI.calc(main_data, second_data, variable)
}
\arguments{
\item{main_data}{A data set needs to be specified.}

\item{second_data}{A second data set needs to be specified.}

\item{variable}{A variable needs to be specified.}
}
\description{
This function allows the calculation of the SSI for a given variable.
}
\examples{
SSI.calc(data1, data2, "catvar")
}
\keyword{creditR}
