% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.R
\docType{methods}
\name{summary,CDS-method}
\alias{summary,CDS-method}
\title{\code{summary} method displays only the essential info about the CDS
class object.}
\usage{
\S4method{summary}{CDS}(object, ...)
}
\description{
\code{summary} method displays only the essential info about the CDS
class object.
}

