% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/summary.R
\docType{methods}
\name{summary}
\alias{summary}
\alias{summary,CDS-method}
\title{Summary Method}
\usage{
\S4method{summary}{CDS}(object, ...)
}
\arguments{
\item{object}{the input \code{CDS} object}

\item{...}{additional arguments to pass in}
}
\description{
\code{summary} method displays only the essential info about the CDS
class object.
}

