% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.default.flag.R
\name{summary.default.flag}
\alias{summary.default.flag}
\title{Summary by Default Flag}
\usage{
\method{summary}{default.flag}(data, default_flag, variable)
}
\arguments{
\item{data}{A data set needs to be specified.}

\item{default_flag}{The default flag need to specified.}

\item{variable}{A variable need to be specified to calculate summary statistics.}
}
\description{
This function allows calculation of the summary statistics of a given variable based on good and bad observations.
}
\examples{
default_f <- c('1','0','0', '1')
birth_year <- c(1980, 1985, 1971, 1990)
salary <- c(20000,NA,10000, 10050)
example_data <- data.frame(default_f,birth_year,salary)
summary.default.flag(example_data, "default_f", "birth_year")
}
\keyword{creditR}
