% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{supersection_multipliers}
\alias{supersection_multipliers}
\title{supersection_multipliers}
\format{
A tibble with 88 rows and 3 variables:
\describe{
\item{supersection}{USFS supersection}
\item{baseline_emission_multipliers}{supersection-associated multiplier}
\item{project_emission_multipliers}{supersection-associated multiplier}
}
}
\source{
data-raw/hwp.R
}
\usage{
supersection_multipliers
}
\description{
USFS supersection-associated emission multipliers
}
\keyword{datasets}
