\name{testDate}
\alias{testDate}
\title{Function for testing the date entered}
\usage{
testDate(date)
}
\arguments{
  \item{date}{date entered}
}
\value{
TRUE if date is valid, FALSE if date is in future, and
"Invalid date format. Must be YYYY-MM-DD" if format is
wrong
}
\description{
Function for testing the date entered
}

