% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_test_balanced_split.R
\name{train_test_balanced_split}
\alias{train_test_balanced_split}
\title{Train and Test Balanced Split}
\usage{
train_test_balanced_split(data, default_flag, balance_count,
  seed_value = 1, ratio)
}
\arguments{
\item{data}{A data set needs to be specified.}

\item{default_flag}{The default flag need to specified as string.}

\item{balance_count}{Desired number of good ar bad observations need to specified.}

\item{seed_value}{A seed value need to be specified for replicability.}

\item{ratio}{The percentage of train data should be specified.}
}
\description{
This function allows you to separate your data set in a balanced manner as train and test.
}
\examples{
default_f <- c('1','0','0', '1','1','0','0','1','1')
birth_year <- c(1980, 1985, 1971,1971,1985,1971,1980,1980,1985)
PD <- c(0.1,0.12, 0.2, 0.23, 0.28,0.33,0.39,0.45, O.54)
example_data <- data.frame(default_f,birth_year,PD)
train_test_balanced_split(example_data,default_flag ="default_f", balance_count = 2,seed_value = 1, ratio = 0.90 )
}
\keyword{creditR}
