% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{update,CDS-method}
\alias{update,CDS-method}
\title{Update with New CDS}
\usage{
\S4method{update}{CDS}(object, isPriceClean = FALSE, spread = NULL, ...)
}
\arguments{
\item{object}{is a \code{CDS} class object.}

\item{spread}{is the new spread in bps.}

\item{isPriceClean}{specifies whether it is a dirty or clean upfront.}

\item{...}{Additional arguments to the call, or arguments with
   changed values. Use \code{name = NULL} to remove the argument \code{name}.}
}
\value{
a \code{CDS} class object
}
\description{
\code{update} spread or ptsUpfront or upfront based on a new CDS class object.
}
\examples{
\dontrun{
## build a CDS class object
cds1 <- CDS(date = as.Date("2014-05-07"), tenor = 5, spread = 50, coupon = 100)

## update
update(cds1, spread = 55)
}
}

