% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable.clustering.gini.R
\name{variable.clustering.gini}
\alias{variable.clustering.gini}
\title{Variable Clustering Gini}
\usage{
variable.clustering.gini(data, default_flag, cluster_number = "optimal")
}
\arguments{
\item{data}{A data set needs to be defined.}

\item{default_flag}{Default flag must be specified as a string.}

\item{cluster_number}{(Default = "optimal"), If the optimal option is selected, clustering is performed over the optimal number of clustering determined by function.}
}
\description{
This function allows you to perform variable clustering and calculate the gini values for a given data set. Number of clusters can be determined optimally or manually.The clusGap function is used for optimal clustering.
}
\examples{
variable.clustering.gini(credit_data, "default_flag", cluster_number = "optimal")
}
\keyword{creditR}
