% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vif.calc.R
\name{vif.calc}
\alias{vif.calc}
\title{Variance Inflation Factor}
\usage{
vif.calc(mod)
}
\arguments{
\item{mod}{A model needs to be specified.}
}
\description{
This function allows the calculation VIF for a given model.
}
\examples{
default_f <- c('1','0','0', '1','1','0','0','1','1')
birth_year <- c(1980, 1985, 1971,1971,1985,1971,1980,1980,1985)
job <- c(1,1,2, 2,2,3,3,2,3)
example_data <- data.frame(default_f,birth_year,job)
vif.calc(glm( formula  = default_f~., family = binomial(link = "logit"), data = example_data))
}
\keyword{creditR}
