% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woe.get.clear.data.R
\name{woe.get.clear.data}
\alias{woe.get.clear.data}
\title{Get Clear Data After WOE binning}
\usage{
woe.get.clear.data(data, default_flag, prefix)
}
\arguments{
\item{data}{A data set needs to be defined.}

\item{default_flag}{Default flag must be specified as a string.}

\item{prefix}{Prefix must be specified as a string. (e.g. ("woe."))}
}
\description{
This function allows you to get a clear data after using woe.binning.deploy function.
}
\examples{
woe.get.clear.data(example_data, "default_f","woe.")
}
\keyword{creditR}
