% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woe.glm.feature.importance.R
\name{woe.glm.feature.importance}
\alias{woe.glm.feature.importance}
\title{Woe Glm Feature Importance}
\usage{
woe.glm.feature.importance(model_data, model, default_flag)
}
\arguments{
\item{model_data}{A model data set needs to be specified.}

\item{model}{A GLM model built by WOE transformation should be given.}

\item{default_flag}{The default flag need to specified as string.}
}
\description{
This function allows the calculate feature importance for a given glm model with WOE transformation method.
}
\examples{
woe.glm.feature.importance(model_data_v1, my_model,"default_f")
}
\keyword{creditR}
