% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woe.table.calc.R
\name{woe.table.calc}
\alias{woe.table.calc}
\title{Calculate WOE Table}
\usage{
woe.table.calc(data, default_flag, variable)
}
\arguments{
\item{data}{A data set needs to be specified.}

\item{default_flag}{The default flag need to specified as string.}

\item{variable}{A variable need to be defined as a string.}
}
\description{
This function allows the calculation of the WOE table for a given variable. It can be useful for manual interventions in WOE transformations.
}
\examples{
default_f <- c('1','0','0', '1','1','0','0','1')
birth_year <- c(1980, 1985, 1971,1971,1985,1971,1980,1980)
salary <- c(20000,2000,10000, 10050,2000,21500,12422,10521)
example_data <- data.frame(default_f,birth_year,salary)
woe.table.calc(example_data, "default_f", "birth_year")
}
\keyword{packagename}
