% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adj.next.bus.day.R
\name{adj.next.bus.day}
\alias{adj.next.bus.day}
\title{Adjust to next business day.}
\usage{
adj.next.bus.day(date)
}
\arguments{
\item{date,}{a \code{Date} type.}
}
\value{
Date adjusted to the following business day
}
\description{
\code{adj.next.bus.day} gets the next business day following 5D bus day
convention.
}

