% Generated by roxygen2 (4.0.1): do not edit by hand
\name{CDS}
\alias{CDS}
\title{Build a \code{CDS} class object given the input about a CDS
contract.}
\usage{
CDS(contract = "SNAC", name = NULL, RED = NULL, date = Sys.Date(),
  baseDate = as.Date(date) + 2, currency = "USD", types = NULL,
  rates = NULL, expiries = NULL, valueDate = NULL, benchmarkDate = NULL,
  startDate = NULL, endDate = NULL, stepinDate = NULL, maturity = NULL,
  tenor = NULL, spread = NULL, coupon = 100, recovery.rate = 0.4,
  upfront = NULL, ptsUpfront = NULL, isPriceClean = FALSE,
  notional = 1e+07, payAccruedOnDefault = TRUE,
  convention = as.vector(data.frame(mmDCC = "ACT/360", calendar = "None",
  fixedSwapDCC = "30/360", floatSwapDCC = "ACT/360", fixedSwapFreq = "6M",
  floatSwapFreq = "3M", holidays = "None", dccCDS = "ACT/360", badDayConvCDS =
  "F", badDayConvZC = "M")))
}
\arguments{
\item{contract}{is the contract type, default SNAC}

\item{name}{is the name of the reference entity. Optional.}

\item{RED}{alphanumeric code assigned to the reference entity. Optional.}

\item{date}{is when the trade is executed, denoted as T. Default
is \code{Sys.Date}. The date format should be in "YYYY-MM-DD".}

\item{baseDate}{is the start date for the IR curve. Default is date + 2 weekdays.
Format must be YYYY-MM-DD.}

\item{currency}{in which CDS is denominated.}

\item{types}{is a string indicating the names of the instruments
used for the yield curve. 'M' means money market rate; 'S' is swap
rate.}

\item{rates}{is an array of numeric values indicating the rate of
each instrument.}

\item{expiries}{is an array of characters indicating the maturity
of each instrument.}

\item{valueDate}{is the date for which the present value of the CDS
is calculated. aka cash-settle date. The default is T + 3.}

\item{benchmarkDate}{Accrual begin date.}

\item{startDate}{is when the CDS nomially starts in terms of
premium payments, i.e. the number of days in the first period (and
thus the amount of the first premium payment) is counted from this
date. aka accrual begin date.}

\item{endDate}{aka maturity date. This is when the contract expires
and protection ends. Any default after this date does not trigger a
payment.}

\item{stepinDate}{default is T + 1.}

\item{maturity}{date of the CDS contract.}

\item{tenor}{of contract. By default is set as 5}

\item{spread}{CDS par spread in bps.}

\item{coupon}{quoted in bps. It specifies the payment amount from
the protection buyer to the seller on a regular basis. The default
is 100 bps.}

\item{recovery.rate}{in decimal. Default is 0.4.}

\item{upfront}{is quoted in the currency amount. Since a standard
contract is traded with fixed coupons, upfront payment is
introduced to reconcile the difference in contract value due to the
difference between the fixed coupon and the conventional par
spread. There are two types of upfront, dirty and clean. Dirty
upfront, a.k.a. Cash Settlement Amount, refers to the market value
of a CDS contract. Clean upfront is dirty upfront less any accrued
interest payment, and is also called the Principal.}

\item{ptsUpfront}{is quoted as a percentage of the notional
amount. They represent the upfront payment excluding the accrual
payment. High Yield (HY) CDS contracts are often quoted in points
upfront. The protection buyer pays the upfront payment if points
upfront are positive, and the buyer is paid by the seller if the
points are negative.}

\item{isPriceClean}{refers to the type of upfront calculated. It is
boolean. When \code{TRUE}, calculate principal only. When
\code{FALSE}, calculate principal + accrual.}

\item{notional}{is the amount of the underlying asset on which the
payments are based. Default is 1e7, i.e. 10MM.}

\item{payAccruedOnDefault}{is a partial payment of the premium made
to the protection seller in the event of a default. Default is
\code{TRUE}.}

\item{convention}{a named vector which contains all the 12 conventional
parameters: mmDCC, calendar, fixedSwapDCC, floatSwapDCC, fixedSwapFreq,
floatSwapFreq, holidays, dccCDS, badDayConvCDS,
and badDayConvZC with their default values}
}
\value{
a \code{CDS} class object including the input informtion on
the contract as well as the valuation results of the contract.
}
\description{
Build a \code{CDS} class object given the input about a CDS
contract.
}
\examples{
# Build a simple CDS class object
require(CDS)
cds <- CDS(date = as.Date("2014-05-07"), tenor = 5, spread = 50, coupon = 100)
}

