% Generated by roxygen2 (4.0.1): do not edit by hand
\name{checkDate}
\alias{checkDate}
\title{Function for testing the date entered}
\usage{
checkDate(date)
}
\arguments{
\item{date}{date entered}
}
\value{
TRUE if date is valid, FALSE if date is in future, and
"Invalid date format. Must be YYYY-MM-DD" if format is wrong
}
\description{
Function for testing the date entered
}

