% Generated by roxygen2 (4.0.1): do not edit by hand
\name{check.inputs}
\alias{check.inputs}
\title{Check whether inputs from the data frame are valid}
\usage{
check.inputs(x, date.var = "date", currency.var = "currency",
  maturity.var = "maturity", tenor.var = "tenor", spread.var = "spread",
  coupon.var = "coupon", notional.var = "notional",
  RR.var = "recovery.rate")
}
\arguments{
\item{spread.var}{name of column in x containing  spreads in bps.}

\item{date.var}{column in x containing the date to be used for pricing and
  other calculations.}

\item{x}{data frame, contains all the relevant columns.}

\item{currency.var}{character, column in x containing currency.}

\item{maturity.var}{character, column in x containing maturity date.}

\item{tenor.var}{character, column in x containing tenors.}

\item{coupon.var}{character, column in x containing coupon rates in basis
points. It specifies the payment amount from the protection buyer to the
seller on an annual basis.}

\item{notional.var}{character, column in x containing the amount of the
  underlying asset on which the payments are based.}

\item{RR.var}{character, column in x containing recovery rates. ISDA model
standard recovery rate asscumption is 0.4.}
}
\value{
a data frame if not stopped by errors.
}
\description{
\code{check.inputs} checks whether a data frame's inputs are valid. It is a
minimum set of checks. Things such as recovery.rate var are not checked,
because some functions don't need them as input.
}

