% Generated by roxygen2 (4.0.1): do not edit by hand
\name{check.inputs}
\alias{check.inputs}
\title{Check whether inputs from the data frame are valid}
\usage{
check.inputs(x, date.var = "date", currency.var = "currency",
  maturity.var = "maturity", tenor.var = "tenor", spread.var = "spread",
  coupon.var = "coupon", notional.var = "notional")
}
\arguments{
\item{spread.var}{name of column in x containing  spreads in bps.}

\item{date.var}{column in x containing dates when the trade
is executed, denoted as T. Default is \code{Sys.Date}  + 2 weekdays.}

\item{x}{is the data frame containing all the relevant columns.}

\item{currency.var}{name of column in x containing currencies.}

\item{maturity.var}{name of column in x containing maturity dates.}

\item{tenor.var}{name of column in x containing tenors.}

\item{coupon.var}{name of column in x containing coupon rates in bps.
It specifies the payment amount from the protection buyer to the seller
on a regular basis.}

\item{notional.var}{name of column in x containing the amount of
   the underlying asset on which the payments are based.
   Default is 1e7, i.e. 10MM.}
}
\value{
a data frame if not stopped by errors.
}
\description{
\code{check.inputs} checks whether a data frame's inputs are valid. It
     is a minimum set of checks. Things such as recovery.rate var are
     not checked, because some functions don't need them as input.
}

