% Generated by roxygen2 (4.0.1): do not edit by hand
\name{check.rates.dates}
\alias{check.rates.dates}
\title{Make sure dates are available}
\usage{
check.rates.dates(x, rates, date.var = "date", rate.var = "rates")
}
\arguments{
\item{x}{data frame which contains rates for a specific date}

\item{rates}{data frame which contains rates and dates}

\item{date.var}{character which indicates the name of date column}

\item{rate.var}{character which indicates the name of rate column}
}
\value{
vector containing true or false to indicate whether
that specific trade date in X is contained in rates
}
\description{
\code{check.rates.dates.R} makes sure the dates in data frame x
being used by the upfrontdf function are also present in the
rates data frame being entered into the upfrontdf function (if
it is not contained in the rates data frame, the upfrontdf
function will be unable to return the upfront value)
match with the rates on markit.com for that specific date.
}

