% Generated by roxygen2 (4.0.1): do not edit by hand
\name{CS10.DF}
\alias{CS10.DF}
\title{function that takes a vector of CDS objects to return a vector of CS 10 calculations for each object}
\usage{
CS10.DF(cds)
}
\arguments{
\item{cds}{vector of CDS objects}
}
\value{
vector of CS10 values
}
\description{
function that takes a vector of CDS objects to return a vector of CS 10 calculations for each object
}

