% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get.raw.markit}
\alias{get.raw.markit}
\title{Get raw data from Markit website.}
\usage{
get.raw.markit(date, currency)
}
\arguments{
\item{URL}{is the link containing the rates.}

\item{verbose}{option. Default \code{FALSE}.}
}
\value{
a xml file crawled from the \code{URL}.
}
\description{
\code{get.raw.markit} downloads the rates zip file from Markit website,
unzips and parses the XML
}

