% Generated by roxygen2 (4.0.1): do not edit by hand
\name{JPY.condition}
\alias{JPY.condition}
\title{Japanese Dates}
\usage{
JPY.condition(currency = currency, TDate = Tdate, baseDate = baseDate)
}
\arguments{
\item{currency}{in which CDS is denominated.}

\item{Tdate}{is when the trade is executed, denoted as T. Default
is \code{Sys.Date}. The date format should be in "YYYY-MM-DD".}

\item{baseDate}{is the start date for the IR curve. Default is TDate + 2 weekdays.
Format must be YYYY-MM-DD.}
}
\value{
updated \code{baseDate}
}
\description{
\code{JPY.condition} examines whether the currency is JPY, and adjusts base
   date accordingly if so
}

