% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pd.spread}
\alias{pd.spread}
\title{function to calculate spread using the probability of default, time and recovery rate}
\usage{
pd.spread(PD, time, Recovery = 0.4)
}
\arguments{
\item{R}{recovery rate in decimals}

\item{PD}{probability of default in decimals}

\item{time}{in years}
}
\value{
spread in basis points
}
\description{
function to calculate spread using the probability of default, time and recovery rate
}

