% Generated by roxygen2 (4.0.1): do not edit by hand
\name{PV01}
\alias{PV.01}
\alias{PV01}
\title{Calculate PV01}
\usage{
PV.01(x, principal.var = "principal", spread.var = "spread",
  coupon.var = "coupon", notional.var = "notional")
}
\arguments{
\item{x}{data frame containing the principal, parSpread, coupon and notional}

\item{principal.var}{name of the column containing the principal or
clean upfront values of the CDS}

\item{spread.var}{name of the column containing the CDS contract in basis points}

\item{coupon.var}{name of the column containing the CDS contract in basis points}

\item{notional.var}{name of the column containing the notional amount (in currency terms).}
}
\value{
Vector containing the PV01 values using the formula:
(principal/notional)*(10000/(parSpread-coupon))
}
\description{
\code{PV01} to calculate present value 01 or present value of a stream of 1bp payments
}

