% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rec.risk.01}
\alias{rec.risk.01}
\title{Calculate RR change}
\usage{
rec.risk.01(x, date.var = "date", currency.var = "currency",
  maturity.var = "maturity", tenor.var = "tenor", spread.var = "spread",
  coupon.var = "coupon", RR.var = "recovery.rate",
  notional.var = "notional")
}
\arguments{
\item{x}{is the data frame containing all the relevant columns.}

\item{date.var}{name of column in x containing dates when the trade
is executed, denoted as T. Default is \code{Sys.Date}  + 2 weekdays.}

\item{currency.var}{name of column in x containing currencies.}

\item{maturity.var}{name of column in x containing maturity dates.}

\item{tenor.var}{name of column in x containing tenors.}

\item{spread.var}{name of column in x containing  par spreads in bps.}

\item{coupon.var}{name of column in x containing coupon rates in bps.
It specifies the payment amount from the protection buyer to the seller
on a regular basis.}

\item{RR.var}{name of column in x containing recovery
rates in decimal.}

\item{notional.var}{name of column in x containing the amount of
   the underlying asset on which the payments are based.
   Default is 1e7, i.e. 10MM.}
}
\value{
a vector containing the change in upfront when there is a 1
percent increase in recovery rate, for each corresponding CDS contract.
}
\description{
\code{rec.risk.01} calculate the amount of change in upfront when there is a 1%
increase in recovery rate for a data frame of CDS contracts.
}
\examples{
x <- data.frame(date = c(as.Date("2014-04-22"), as.Date("2014-04-22")),
currency = c("USD", "EUR"),
tenor = c(5, 5),
spread = c(120, 110),
coupon = c(100, 100),
recovery.rate = c(0.4, 0.4),
notional = c(1e7, 1e7))
result <- rec.risk.01(x)
}

