% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,CDS-method}
\title{Show Method}
\usage{
\S4method{show}{CDS}(object)
}
\arguments{
\item{object}{the input \code{CDS} object}

\item{...}{additional arguments to pass in}
}
\description{
\code{show} shows a CDS class object.
}

