% Generated by roxygen2 (4.0.1): do not edit by hand
\name{spread}
\alias{spread}
\title{\code{spread} to calculate conventional spread using the upfront or ptsUpfront values}
\usage{
spread(x, currency.var = "currency", date.var = "date",
  coupon.var = "coupon", tenor.var = "tenor", maturity.var = "maturity",
  RR.var = "recovery.rate", upfront.var = "upfront",
  ptsUpfront.var = "ptsUpfront", isPriceClean = FALSE, notional = 1e+07,
  freqCDS = "Q", stubCDS = "F", payAccruedAtStart = FALSE,
  payAccruedOnDefault = TRUE)
}
\arguments{
\item{ptsUpfront.var}{}

\item{isPriceClean}{}

\item{notional}{}

\item{freqCDS}{date interval of the CDS contract.}

\item{stubCDS}{is a character indicating the presence of a stub.}

\item{payAccruedAtStart}{whether pay at start date the accrual amount}

\item{payAccruedOnDefault}{}

\item{x}{data frame, contains all the relevant columns.}

\item{currency.var}{character, column in x containing currency.}

\item{date.var}{character, column in x containing date variable.}

\item{coupon.var}{character, column in x containing coupon rates in basis
points. It specifies the payment amount from the protection buyer to the
seller on an annual basis.}

\item{tenor.var}{character, column in x containing tenors.}

\item{maturity.var}{character, column in x containing maturity date.}

\item{RR.var}{character, column in x containing recovery rates. ISDA model
standard recovery rate asscumption is 0.4.}
}
\value{
a numeric indicating the spread.
}
\description{
\code{spread} to calculate conventional spread using the upfront or ptsUpfront values
}

