% Generated by roxygen2 (4.0.1): do not edit by hand
\name{spread}
\alias{spread}
\title{\code{spread} to calculate conventional spread using the upfront or ptsUpfront values}
\usage{
spread(date, baseDate = as.Date(date) + 2, currency = "USD", types = NULL,
  rates = NULL, expiries = NULL, mmDCC = "ACT/360",
  fixedSwapFreq = "6M", floatSwapFreq = "3M", fixedSwapDCC = "30/360",
  floatSwapDCC = "ACT/360", badDayConvZC = "M", holidays = "None",
  valueDate = NULL, benchmarkDate = NULL, startDate = NULL,
  endDate = NULL, stepinDate = NULL, maturity = NULL, tenor = NULL,
  dccCDS = "ACT/360", freqCDS = "Q", stubCDS = "F", badDayConvCDS = "F",
  calendar = "None", upfront = NULL, ptsUpfront = NULL, coupon = 100,
  recovery.rate = 0.4, payAccruedAtStart = FALSE, notional = 1e+07,
  payAccruedOnDefault = TRUE)
}
\arguments{
\item{types}{is the type of interest rate that we are using in CDS calculation}

\item{rates}{is the interest rate we are using for calculating a spread.
notice that when we trade CDS on a trade date, we don't use the
interest rate on that date to calculate the pricing; instead, we
use the interest rate of the previous business day. This is
set by the ISDA Standard Model}

\item{expiries}{is the type of CDS expiries we use to calculate. Expiries
are usually different among different currencies.}

\item{payAccruedAtStart}{whether pay at start date the accrual amount}

\item{spread}{in bps}

\item{mmDCC}{is the day count convention of the instruments.}

\item{fixedSwapFreq}{is the frequency of the fixed rate of swap
being paid.}

\item{floatSwapFreq}{is the frequency of the floating rate of swap
being paid.}

\item{fixedSwapDCC}{is the day count convention of the fixed leg.}

\item{floatSwapDCC}{is the day count convention of the floating leg.}

\item{badDayConvZC}{is a character indicating how non-business days
are converted.}

\item{holidays}{is an input for holiday files to adjust to business
days.}

\item{valueDate}{T+3 business days}

\item{benchmarkDate}{accrual begin date}

\item{startDate}{start date}

\item{endDate}{end date}

\item{stepinDate}{T+1}

\item{dccCDS}{day count convention of the CDS. Default is ACT/360.}

\item{freqCDS}{date interval of the CDS contract.}

\item{stubCDS}{is a character indicating the presence of a stub.}

\item{badDayConvCDS}{refers to the bay day conversion for the CDS
coupon payments. Default is "F", following.}

\item{calendar}{refers to any calendar adjustment for the CDS.}

\item{date}{is when the trade is executed, denoted as T. Default
is \code{Sys.Date}. The date format should be in "YYYY-MM-DD".}

\item{currency}{in which CDS is denominated.}

\item{maturity}{date of the CDS contract.}

\item{tenor}{of contract. By default is set as 5}

\item{upfront}{is quoted in the currency amount. Since a standard
contract is traded with fixed coupons, upfront payment is
introduced to reconcile the difference in contract value due to the
difference between the fixed coupon and the conventional par
spread. There are two types of upfront, dirty and clean. Dirty
upfront, a.k.a. Cash Settlement Amount, refers to the market value
of a CDS contract. Clean upfront is dirty upfront less any accrued
interest payment, and is also called the Principal.}

\item{ptsUpfront}{is quoted as a percentage of the notional
amount. They represent the upfront payment excluding the accrual
payment. High Yield (HY) CDS contracts are often quoted in points
upfront. The protection buyer pays the upfront payment if points
upfront are positive, and the buyer is paid by the seller if the
points are negative.}

\item{coupon}{quoted in bps. It specifies the payment amount from
the protection buyer to the seller on a regular basis. The default
is 100 bps.}

\item{recovery.rate}{in decimal. Default is 0.4.}

\item{notional}{is the amount of the underlying asset on which the
payments are based. Default is 1e7, i.e. 10MM.}

\item{payAccruedOnDefault}{is a partial payment of the premium made
to the protection seller in the event of a default. Default is
\code{TRUE}.}
}
\value{
a numeric indicating the spread.
}
\description{
\code{spread} to calculate conventional spread using the upfront or ptsUpfront values
}

