% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{summary,CDS-method}
\alias{summary,CDS-method}
\title{\code{summary} method displays only the essential info about the CDS
class object.}
\usage{
\S4method{summary}{CDS}(object, ...)
}
\arguments{
\item{object}{an object for which a summary is desired.}

\item{...}{additional arguments affecting the summary produced.}
}
\description{
\code{summary} method displays only the essential info about the CDS
class object.
}

