% Generated by roxygen2 (4.0.1): do not edit by hand
\name{upfront}
\alias{upfront}
\title{Calculate Upfront Payments}
\usage{
upfront(x, rates, currency.var = "currency", notional = 1e+07,
  date.var = "date", spread.var = "spread", coupon.var = "coupon",
  tenor.var = "tenor", maturity.var = "maturity",
  recovery.var = "recovery", isPriceClean = FALSE)
}
\arguments{
\item{rates}{is an array of numeric values indicating the rate of
each instrument.}

\item{notional}{is the amount of the underlying asset on which the
payments are based. Default is 1e7, i.e. 10MM.}

\item{recovery.var}{f column in x containing recovery
rates in decimal.}

\item{isPriceClean}{refers to the type of upfront calculated. It is
       boolean. When \code{TRUE}, calculate principal only. When
       \code{FALSE}, calculate principal + accrual.}

\item{x}{data frame, contains all the relevant columns.}

\item{currency.var}{character, column in x containing currency.}

\item{date.var}{character, column in x containing date variable.}

\item{spread.var}{character, column in x containing spread in basis points.}

\item{coupon.var}{character, column in x containing coupon rates in basis
points. It specifies the payment amount from the protection buyer to the
seller on an annual basis.}

\item{tenor.var}{character, column in x containing tenors.}

\item{maturity.var}{character, column in x containing maturity date.}
}
\value{
vector of upfront values (with accrual) in the same order
}
\description{
\code{upfront} takes a dataframe of variables on CDSs to return a vector of
upfront values. Note that all CDS in the data frame must be denominated in
the same currency.
}

