% Generated by roxygen2 (4.0.1): do not edit by hand
\name{upfront}
\alias{upfront}
\title{Calculate Upfront Payments}
\usage{
upfront(x, rates, currency.var = "currency", notional = 1e+07,
  date.var = "date", spread.var = "spread", coupon.var = "coupon",
  tenor.var = "tenor", maturity.var = "maturity",
  recovery.var = "recovery", isPriceClean = FALSE)
}
\arguments{
\item{rates}{is an array of numeric values indicating the rate of
each instrument.}

\item{notional}{is the amount of the underlying asset on which the
payments are based. Default is 1e7, i.e. 10MM.}

\item{date.var}{name of column in x containing dates when the trade
is executed, denoted as T. Default is \code{Sys.Date}  + 2 weekdays.}

\item{spread.var}{name of column in x containing  par spreads in bps.}

\item{recovery.var}{f column in x containing recovery
rates in decimal.}

\item{x}{is the data frame containing all the relevant columns.}

\item{currency.var}{name of column in x containing currencies.}

\item{coupon.var}{name of column in x containing coupon rates in bps.
It specifies the payment amount from the protection buyer to the seller
on a regular basis.}

\item{tenor.var}{name of column in x containing tenors.}

\item{maturity.var}{name of column in x containing maturity dates.}

\item{isPriceClean}{refers to the type of upfront calculated. It is
boolean. When \code{TRUE}, calculate principal only. When
\code{FALSE}, calculate principal + accrual.}
}
\value{
vector of upfront values (with accrual) in the same order
}
\description{
\code{upfront} takes a dataframe of variables on CDSs to return a vector of
upfront values. Note that all CDS in the data frame must be denominated in
the same currency.
}

