% Generated by roxygen2 (4.0.1): do not edit by hand
\name{upfrontdf}
\alias{upfrontdf}
\title{\code{upfrontdf} takes a dataframe of variables on CDSs to return a vector of
upfront values. Note that all CDS in the data frame must be denominated in
the same currency.}
\usage{
upfrontdf(x, rates, currency.var = "currency", notional = 10000000,
  date.var = "date", spread.var = "spread", coupon.var = "coupon",
  tenor.var = "tenor", maturity.var = "maturity", isPriceClean = FALSE)
}
\arguments{
\item{x}{dataframe containing variables date.var, spread.var, coupon.var and
maturity.var.}

\item{rates}{dataframe containing dates and rates for those dates. Note that
the date column over here refers to the date for which the adjacent
interest rate curve would apply. So if it says "2014-07-25", the interest
rate curve is from "2014-07-24".}

\item{currency.var}{name of the column for currencies in the dataframe.}

\item{notional}{values of CDSs in the dataframe. Defualt is 10 million.}

\item{date.var}{name of the column containing dates. By default is "date"}

\item{spread.var}{name of the column containing spreads. By default is
"spread".}

\item{coupon.var}{name of the column containing the coupon rates. By default
is "coupon"}

\item{maturity.var}{name of the column containing the maturity dates (note:
this is different from tenor i.e. it is a proper date like "2019-06-20" and
not "5Y"). By default is "maturity".}

\item{tenor.var}{name of the column containing the tenors of the CDS contracts.
Note that we can only provide either the tenor or the maturity date, not both.}

\item{isPriceClean}{boolean to specify if you want the dirty upfront or the
clean upfront (principal).}
}
\value{
vector of upfront values (with accrual) in the same order
}
\description{
\code{upfrontdf} takes a dataframe of variables on CDSs to return a vector of
upfront values. Note that all CDS in the data frame must be denominated in
the same currency.
}

